% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{add.log.noise}
\alias{add.log.noise}
\title{Add log normal noise to value}
\usage{
add.log.noise(val, noise)
}
\arguments{
\item{val}{value}

\item{noise}{noise}
}
\description{
Add log normal noise to value
}
