% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{apply.hcr}
\alias{apply.hcr}
\title{Apply harvest control rule (HCR)}
\usage{
apply.hcr(HCR = "noF", biomass = NULL, refs = NULL)
}
\arguments{
\item{HCR}{HCR type}

\item{biomass}{Curent biomass (either SSB or ESB) should correspond to biomass type in refs!}

\item{refs}{Reference points, should contain Fmsy, Bpa, Blim}
}
\description{
Apply harvest control rule (HCR)
}
