% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{apply.ms}
\alias{apply.ms}
\title{Apply management stratey (MS)}
\usage{
apply.ms(F, MStype = "noF", SSB = NA, Bpa = NA)
}
\arguments{
\item{F}{HERE:}

\item{MStype}{HERE:}

\item{SSB}{HERE:}

\item{Bpa}{HERE:}
}
\description{
Apply management strategy (MS)
}
