% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calc.prod}
\alias{calc.prod}
\title{Calculate production based on Schaefer}
\usage{
calc.prod(B, r, K)
}
\arguments{
\item{B}{biomass}

\item{r}{intrinsic growth rate}

\item{K}{carrying capacity}
}
\description{
Calculates production based on the Schaefer model
}
