% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calc.prod.thorson}
\alias{calc.prod.thorson}
\title{Calculate production based on Thorson}
\usage{
calc.prod.thorson(B, r, K, n, fmsy)
}
\arguments{
\item{B}{biomass}

\item{r}{intrinsic growth rate}

\item{K}{carrying capacity}
}
\description{
Calculates production based on the Thorson model
}
