% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calc.ssb}
\alias{calc.ssb}
\title{Calculate SSB}
\usage{
calc.ssb(B, F, matpars)
}
\arguments{
\item{B}{biomass}

\item{F}{fishing mortality}

\item{matpars}{maturity parameters}
}
\description{
Calculate SSB
}
