% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{esb2ssb}
\alias{esb2ssb}
\title{Convert ESB to SSB}
\usage{
esb2ssb(ESB, F, pars, exp = TRUE)
}
\arguments{
\item{ESB}{Exploitable stock biomass}

\item{F}{fishing mortality}

\item{pars}{conversion parameters}
}
\value{
Spawning stock biomass (SSB)
}
\description{
Convert ESB to SSB
}
