% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{est.pars.esb2ssb}
\alias{est.pars.esb2ssb}
\title{Estimate parameters for ESB to SSB conversion}
\usage{
est.pars.esb2ssb(data, plot = TRUE, log = FALSE, verbose = TRUE)
}
\arguments{
\item{data}{data frame with a column, ESB exploitable stock biomass, SSB spawning stock biomass, and F fishing mortality}

\item{plot}{Plot the results? Default: TRUE}

\item{log}{Fit ont log scale?}

\item{verbose}{Print info?}
}
\description{
Estimate parameters for ESB to SSB conversion
}
