% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.pars.spict}
\alias{get.pars.spict}
\title{Extract relevant parameters from fitted spict object}
\usage{
get.pars.spict(fit, verbose = TRUE)
}
\arguments{
\item{fit}{Fitted spict object}

\item{verbose}{Print information}
}
\description{
Extract relevant parameters from fitted spict object
}
