% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mackerel}
\alias{mackerel}
\title{Mackerel}
\format{
A dataframe consisting of:
\itemize{
  \item \code{year}  Year
  \item \code{TSB}  Total stock biomass
  \item \code{SSB}  Spawning stock biomass
  \item \code{ESB}  Exploitable stock biomass
  \item \code{catch}  Catches
  \item \code{F}  Fishing mortality
}
}
\source{
Jones, R., 1976. Mesh regulation in the demersal fisheries of the South
China Sea area. Manila, South China Sea Fisheries Development and Coordinating
Programme, SCS/76/WP/34: 75 p.
}
\usage{
data(mackerel)
}
\description{
Mackerel assessment data
}
\examples{
data(mackerel)
head(mackerel)

}
\keyword{data}
\keyword{dataset}
\keyword{mackerel}
