% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.R
\name{perform.ssb}
\alias{perform.ssb}
\title{Performance metrics - spawning stock biomass}
\usage{
perform.ssb(mse)
}
\arguments{
\item{mse}{HERE:}
}
\description{
Calculate spawning stock biomass performance metrics
}
