% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.R
\name{perform.yielddiff}
\alias{perform.yielddiff}
\title{Performance metrics - yield variability}
\usage{
perform.yielddiff(mse)
}
\arguments{
\item{mse}{HERE:}
}
\description{
Calculate yield variability performance metrics
}
