% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.ts.single}
\alias{plot.ts.single}
\title{Plot single time series}
\usage{
\method{plot}{ts.single}(
  mse,
  quant = c("ssbObs", "yImp"),
  col = c("darkorange", "goldenrod2"),
  scenario = 1,
  ms = 1,
  ylab = "Biomass '000t",
  yr = NULL,
  rep = NULL
)
}
\arguments{
\item{mse}{HERE:}

\item{quant}{HERE:}

\item{col}{HERE:}

\item{scenario}{HERE:}

\item{ms}{HERE:}

\item{ylab}{HERE:}

\item{yr}{HERE:}

\item{rep}{HERE:}
}
\description{
Plot single time series
}
