% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotprodmse.explo}
\alias{plotprodmse.explo}
\title{Plot explo}
\usage{
plotprodmse.explo(x, pars, ms = 1)
}
\arguments{
\item{pars}{HERE:}

\item{ms}{HERE:}

\item{mse}{HERE:}
}
\description{
Plot explo
}
