% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{run.explo}
\alias{run.explo}
\title{Run explo}
\usage{
run.explo(
  mse,
  yr = NULL,
  fmsy = seq(0.1, 1, 0.05),
  btrigger = seq(1000, 5000, 200),
  nsim = NULL,
  ncores = 1
)
}
\arguments{
\item{mse}{HERE:}

\item{yr}{HERE:}

\item{fmsy}{HERE:}

\item{btrigger}{HERE:}

\item{nsim}{HERE:}

\item{ncores}{HERE:}
}
\description{
Run explo
}
