% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{run.mse}
\alias{run.mse}
\title{Run MSE}
\usage{
run.mse(pars, mc.cores = 1, verbose = TRUE)
}
\arguments{
\item{pars}{HERE:}

\item{mc.cores}{HERE:}

\item{verbose}{Print information}
}
\description{
Run MSE
}
