% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sim.log.noise}
\alias{sim.log.noise}
\title{Simulate log normal noise}
\usage{
sim.log.noise(sigma, n = 1)
}
\arguments{
\item{sigma}{standard deviation}

\item{n}{Number of observations}
}
\description{
Simulate log normal noise
}
