% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sim.noise}
\alias{sim.noise}
\title{Simulate normal noise}
\usage{
sim.noise(sigma, n = 1)
}
\arguments{
\item{sigma}{standard deviation}

\item{n}{Number of observations}
}
\description{
Simulate normal noise
}
