% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calc.prod}
\alias{calc.prod}
\title{Calculate production}
\usage{
calc.prod(B, pars, type = 2)
}
\arguments{
\item{B}{biomass}

\item{pars}{List with parameters (r, K, n) plus m for Fletcher and fmsy for}

\item{type}{Either 1 for "Schaefer", 2 for  "Fletcher" or 3 for "Thorson". Default is 2 ("Fletcher").}
}
\description{
Calculates production
}
\details{

}
