% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{run.mse}
\alias{run.mse}
\title{Run MSE}
\usage{
run.mse(pars, proyears = 5, nsim = 10, ncores = 1)
}
\arguments{
\item{pars}{HERE:}

\item{proyears}{HERE:}

\item{nsim}{HERE:}

\item{ncores}{HERE:}
}
\description{
Run MSE
}
